keep decimals

+++ hallon.py
@@ -33,6 +33,6 @@ used_pct = int(round(used*100/total))
 daysleft = page.select("p.usage-daysleft")[0].text.split()[0]
 
 print(json.dumps({"total": int(total),
-                  "left": int(left),
-                  "used": int(used_pct),
+                  "left": float(left),
+                  "used": float(used_pct),
                   "daysleft": int(daysleft)}))
